# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Unisphere API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface OkResponse {
  ok: boolean;
}

export interface User {
  id: number;
  username: string;
  name: string;
  email: string;
  major: string;
  avatar: string;
  cover: string;
  bio: string;
  tags: string[];
  followersCount: number;
  followingCount: number;
  postsCount: number;
  createdAt?: string;
}

export interface UpdateUserRequest {
  name?: string;
  major?: string;
  avatar?: string;
  cover?: string;
  bio?: string;
  tags?: string[];
}

export interface UserSuggestion {
  id: number;
  name: string;
  major: string;
  avatar: string;
  isFollowing: boolean;
}

export interface Comment {
  id: number;
  postId: number;
  author: string;
  avatar: string;
  text: string;
  time: string;
}

export interface Post {
  id: number;
  author: string;
  major: string;
  avatar: string;
  content: string;
  hashtags: string;
  image?: string | null;
  likes: number;
  comments: number;
  isLiked: boolean;
  isOwn: boolean;
  commentList: Comment[];
}

export interface CreatePostRequest {
  content: string;
  hashtags?: string;
  image?: string | null;
}

export interface CreateCommentRequest {
  text: string;
}

export interface Society {
  id: number;
  name: string;
  desc: string;
  icon: string;
  members: number;
  joined: boolean;
}

export interface Event {
  id: number;
  name: string;
  date: string;
  location: string;
  image: string;
  category: string;
  attendees: number;
  rsvp: boolean;
}

export interface Job {
  id: number;
  title: string;
  company: string;
  location: string;
  type: string;
  salary: string;
  tags: string[];
  logo: string;
  saved: boolean;
}

export interface ChatContact {
  id: number;
  name: string;
  avatar: string;
  lastMsg: string;
  time: string;
  online: boolean;
  unread: number;
}

export interface Message {
  id: number;
  /** 'sent' or 'received' */
  type: string;
  text?: string;
  isAttachment: boolean;
  image?: string | null;
  filename?: string | null;
}

export interface SendMessageRequest {
  receiverId: number;
  text: string;
}

export interface Notification {
  id: number;
  type: string;
  icon: string;
  color: string;
  user: string;
  action: string;
  detail: string;
  time: string;
  avatar: string;
  read: boolean;
}

export interface Story {
  id: number;
  name: string;
  image?: string | null;
  isAdd?: boolean;
  isOwn?: boolean;
  createdAt?: string;
}

export interface Building {
  id: number;
  name: string;
  hours: string;
  icon: string;
  x: number;
  y: number;
  color: string;
  mapsUrl: string;
}

export interface TrendingTopic {
  rank: number;
  title: string;
  posts: string;
}

export interface CreateStoryRequest {
  name?: string;
  image: string;
}

export interface SearchResults {
  users: UserSuggestion[];
  posts: Post[];
  societies: Society[];
}

export interface UploadUrlRequest {
  /**
     * Original file name.
     * @minLength 1
     */
  name: string;
  /**
     * File size in bytes.
     * @minimum 1
     */
  size: number;
  /**
     * MIME type of the file (e.g. `image/jpeg`).
     * @minLength 1
     */
  contentType: string;
}

export interface UploadUrlResponse {
  /** Presigned GCS URL for PUT upload. */
  uploadURL: string;
  /** Normalized object path (e.g. `/objects/uploads/uuid`). Store this in your database. */
  objectPath: string;
  metadata?: UploadUrlRequest;
}

export interface ErrorEnvelope {
  error: string;
}

export type SearchParams = {
q: string;
};