# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Unisphere API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /stories/{id}:
    delete:
      operationId: deleteStory
      tags: [misc]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema: { $ref: "#/components/schemas/OkResponse" }
...
  /events/{id}/rsvp:
    post:
      operationId: rsvpEvent
      tags: [events]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Event" }
    delete:
      operationId: unrsvpEvent
      tags: [events]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Event" }
...
  /jobs/{id}/save:
    post:
      operationId: saveJob
      tags: [jobs]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Job" }
    delete:
      operationId: unsaveJob
      tags: [jobs]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Job" }
...
  /messages/{userId}:
    get:
      operationId: getMessages
      tags: [messages]
      parameters:
        - { name: userId, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: Messages with user
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Message" }
  /messages:
    post:
      operationId: sendMessage
      tags: [messages]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SendMessageRequest" }
      responses:
        "201":
          description: Sent
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Message" }
...
  /stories:
    get:
      operationId: getStories
      tags: [misc]
      responses:
        "200":
          description: Stories
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Story" }
    post:
      operationId: createStory
      tags: [misc]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateStoryRequest" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Story" }
...
components:
  schemas:
...
    OkResponse:
      type: object
      required: [ok]
      properties:
        ok: { type: boolean }
...
    Event:
      type: object
      required: [id, name, date, location, image, category, attendees, rsvp]
      properties:
        id: { type: integer }
        name: { type: string }
        date: { type: string }
        location: { type: string }
        image: { type: string }
        category: { type: string }
        attendees: { type: integer }
        rsvp: { type: boolean }
    Job:
      type: object
      required: [id, title, company, location, type, salary, tags, logo, saved]
      properties:
        id: { type: integer }
        title: { type: string }
        company: { type: string }
        location: { type: string }
        type: { type: string }
        salary: { type: string }
        tags:
          type: array
          items: { type: string }
        logo: { type: string }
        saved: { type: boolean }
...
    Message:
      type: object
      required: [id, type, isAttachment]
      properties:
        id: { type: integer }
        type: { type: string, description: "'sent' or 'received'" }
        text: { type: string }
        isAttachment: { type: boolean }
        image: { type: string, nullable: true }
        filename: { type: string, nullable: true }
    SendMessageRequest:
      type: object
      required: [receiverId, text]
      properties:
        receiverId: { type: integer }
        text: { type: string }
...
    Story:
      type: object
      required: [id, name]
      properties:
        id: { type: integer }
        name: { type: string }
        image: { type: string, nullable: true }
        isAdd: { type: boolean }
        isOwn: { type: boolean }
        createdAt: { type: string }
...
    CreateStoryRequest:
      type: object
      required: [image]
      properties:
        name: { type: string }
        image: { type: string }
...