# lib/api-zod/src/generated/types/story.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Unisphere API
 * OpenAPI spec version: 0.1.0
 */

export interface Story {
  id: number;
  name: string;
  image?: string | null;
  isAdd?: boolean;
  isOwn?: boolean;
  createdAt?: string;
}